Move unit tests into its own file
