+++ src/math/atan.rs
@@ -135,51 +135,3 @@ pub fn atan(x: f64) -> f64 {
         z
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::atan;
-    use core::f64;
-
-    #[test]
-    fn sanity_check() {
-        for (input, answer) in [
-            (3.0_f64.sqrt() / 3.0, f64::consts::FRAC_PI_6),
-            (1.0, f64::consts::FRAC_PI_4),
-            (3.0_f64.sqrt(), f64::consts::FRAC_PI_3),
-            (-3.0_f64.sqrt() / 3.0, -f64::consts::FRAC_PI_6),
-            (-1.0, -f64::consts::FRAC_PI_4),
-            (-3.0_f64.sqrt(), -f64::consts::FRAC_PI_3),
-        ]
-        .iter()
-        {
-            assert!(
-                (atan(*input) - answer) / answer < 1e-5,
-                "\natan({:.4}/16) = {:.4}, actual: {}",
-                input * 16.0,
-                answer,
-                atan(*input)
-            );
-        }
-    }
-
-    #[test]
-    fn zero() {
-        assert_eq!(atan(0.0), 0.0);
-    }
-
-    #[test]
-    fn infinity() {
-        assert_eq!(atan(f64::INFINITY), f64::consts::FRAC_PI_2);
-    }
-
-    #[test]
-    fn minus_infinity() {
-        assert_eq!(atan(f64::NEG_INFINITY), -f64::consts::FRAC_PI_2);
-    }
-
-    #[test]
-    fn nan() {
-        assert!(atan(f64::NAN).is_nan());
-    }
-}
