+++ src/math/remquo.rs
@@ -96,14 +96,3 @@ pub fn remquo(mut x: f64, mut y: f64) -> (f64, i32) {
         (x, quo)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::remquo;
-
-    #[test]
-    fn test_q_overflow() {
-        // 0xc000000000000001, 0x04c0000000000004
-        let _ = remquo(-2.0000000000000004, 8.406091369059082e-286);
-    }
-}
