+++ src/math/round.rs
@@ -35,13 +35,3 @@ pub fn round(mut x: f64) -> f64 {
         y
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::round;
-
-    #[test]
-    fn negative_zero() {
-        assert_eq!(round(-0.0_f64).to_bits(), (-0.0_f64).to_bits());
-    }
-}
